+++ CHANGELOG.md
@@ -4,11 +4,11 @@
 
 - Implement new breakpoints and screen sizes
 - Removed `SASS` support 
+- Fixed: Modal closes on drag and release cursor outside modal (#152)
 
 ## 0.7.1
 
 - Fixed label ellipsis shown at incorrect position
-- Fixed: Modal closes on drag and release cursor outside modal (#152)
 
 ## 0.7.0